Changing the structure of the "uploadable" concern to be a base class called "GenericFilesController" for inheritance rather than bringing in the methods by inclusion.  I never actually use the GenericFilesController directly in my code, but in the future Cassiano might do so.
