+++ stash/stash_engine/app/controllers/stash_engine/software_files_controller.rb
@@ -2,17 +2,12 @@ require_dependency 'stash_engine/application_controller'
 require 'stash/url_translator'
 
 module StashEngine
-  class SoftwareFilesController < ApplicationController
+  class SoftwareFilesController < GenericFilesController
 
     def setup_class_info
       @file_model = StashEngine::SoftwareFile
       @resource_assoc = :software_files
     end
-
-    include StashEngine::Concerns::Uploadable
-
-    # attr_reader :resource
-    helper_method :resource
 
   end
 end
